+++ packages/shield-controller/src/ShieldController.test.ts
@@ -61,7 +61,7 @@ function setup({
   decodeTransactionDataHandler?: jest.MockedFunction<DecodeTransactionDataHandler>;
 } = {}) {
   const backend = createMockBackend();
-  const { messenger, baseMessenger } = createMockMessenger();
+  const { messenger, rootMessenger } = createMockMessenger();
 
   const controller = new ShieldController({
     backend,
@@ -75,7 +75,7 @@ function setup({
   return {
     controller,
     messenger,
-    baseMessenger,
+    rootMessenger,
     backend,
     decodeTransactionDataHandler,
   };
@@ -84,10 +84,10 @@ function setup({
 describe('ShieldController', () => {
   describe('checkCoverage', () => {
     it('should trigger checkCoverage when a new transaction is added', async () => {
-      const { baseMessenger, backend } = setup();
+      const { rootMessenger, messenger, backend } = setup();
       const txMeta = generateMockTxMeta();
-      const coverageResultReceived = setupCoverageResultReceived(baseMessenger);
-      baseMessenger.publish(
+      const coverageResultReceived = setupCoverageResultReceived(messenger);
+      rootMessenger.publish(
         'TransactionController:stateChange',
         { transactions: [txMeta] } as TransactionControllerState,
         undefined as never,
@@ -97,14 +97,14 @@ describe('ShieldController', () => {
     });
 
     it('should tolerate calling start and stop multiple times', async () => {
-      const { backend, baseMessenger, controller } = setup();
+      const { backend, rootMessenger, messenger, controller } = setup();
       controller.stop();
       controller.stop();
       controller.start();
       controller.start();
       const txMeta = generateMockTxMeta();
-      const coverageResultReceived = setupCoverageResultReceived(baseMessenger);
-      baseMessenger.publish(
+      const coverageResultReceived = setupCoverageResultReceived(messenger);
+      rootMessenger.publish(
         'TransactionController:stateChange',
         { transactions: [txMeta] } as TransactionControllerState,
         undefined as never,
@@ -114,11 +114,11 @@ describe('ShieldController', () => {
     });
 
     it('should no longer trigger checkCoverage when controller is stopped', async () => {
-      const { controller, baseMessenger, backend } = setup();
+      const { controller, rootMessenger, backend } = setup();
       controller.stop();
       const txMeta = generateMockTxMeta();
       const coverageResultReceived = new Promise<void>((resolve, reject) => {
-        baseMessenger.subscribe(
+        rootMessenger.subscribe(
           'ShieldController:coverageResultReceived',
           (_coverageResult) => resolve(),
         );
@@ -127,7 +127,7 @@ describe('ShieldController', () => {
           100,
         );
       });
-      baseMessenger.publish(
+      rootMessenger.publish(
         'TransactionController:stateChange',
         { transactions: [txMeta] } as TransactionControllerState,
         undefined as never,
@@ -166,12 +166,12 @@ describe('ShieldController', () => {
     });
 
     it('should check coverage when a transaction is simulated', async () => {
-      const { baseMessenger, backend } = setup();
+      const { rootMessenger, messenger, backend } = setup();
       const txMeta = generateMockTxMeta();
-      const coverageResultReceived = setupCoverageResultReceived(baseMessenger);
+      const coverageResultReceived = setupCoverageResultReceived(messenger);
 
       // Add transaction.
-      baseMessenger.publish(
+      rootMessenger.publish(
         'TransactionController:stateChange',
         { transactions: [txMeta] } as TransactionControllerState,
         undefined as never,
@@ -184,9 +184,8 @@ describe('ShieldController', () => {
       txMeta2.simulationData = {
         tokenBalanceChanges: [],
       };
-      const coverageResultReceived2 =
-        setupCoverageResultReceived(baseMessenger);
-      baseMessenger.publish(
+      const coverageResultReceived2 = setupCoverageResultReceived(messenger);
+      rootMessenger.publish(
         'TransactionController:stateChange',
         { transactions: [txMeta2] } as TransactionControllerState,
         undefined as never,
@@ -201,16 +200,15 @@ describe('ShieldController', () => {
     TX_META_SIMULATION_DATA_MOCKS.forEach(
       ({ description, previousSimulationData, newSimulationData }) => {
         it(`should check coverage when ${description}`, async () => {
-          const { baseMessenger, backend } = setup();
+          const { rootMessenger, messenger, backend } = setup();
           const previousTxMeta = {
             ...generateMockTxMeta(),
             simulationData: previousSimulationData,
           };
-          const coverageResultReceived =
-            setupCoverageResultReceived(baseMessenger);
+          const coverageResultReceived = setupCoverageResultReceived(messenger);
 
           // Add transaction.
-          baseMessenger.publish(
+          rootMessenger.publish(
             'TransactionController:stateChange',
             { transactions: [previousTxMeta] } as TransactionControllerState,
             undefined as never,
@@ -224,8 +222,8 @@ describe('ShieldController', () => {
           const txMeta2 = { ...previousTxMeta };
           txMeta2.simulationData = newSimulationData;
           const coverageResultReceived2 =
-            setupCoverageResultReceived(baseMessenger);
-          baseMessenger.publish(
+            setupCoverageResultReceived(messenger);
+          rootMessenger.publish(
             'TransactionController:stateChange',
             { transactions: [txMeta2] } as TransactionControllerState,
             undefined as never,
@@ -259,14 +257,14 @@ describe('ShieldController', () => {
     const MOCK_SIGNATURE_REQUEST = generateMockSignatureRequest();
 
     it('should check signature coverage', async () => {
-      const { baseMessenger, backend } = setup();
+      const { rootMessenger, backend } = setup();
       const coverageResultReceived = new Promise<void>((resolve) => {
-        baseMessenger.subscribe(
+        rootMessenger.subscribe(
           'ShieldController:coverageResultReceived',
           (_coverageResult) => resolve(),
         );
       });
-      baseMessenger.publish(
+      rootMessenger.publish(
         'SignatureController:stateChange',
         {
           signatureRequests: {
@@ -294,16 +292,16 @@ describe('ShieldController', () => {
         .mockImplementationOnce((_signatureRequest: SignatureRequest) => {
           return MOCK_NORMALIZED_SIGNATURE_REQUEST;
         });
-      const { baseMessenger, backend } = setup({
+      const { rootMessenger, backend } = setup({
         normalizeSignatureRequest: normalizeSignatureRequestMock,
       });
       const coverageResultReceived = new Promise<void>((resolve) => {
-        baseMessenger.subscribe(
+        rootMessenger.subscribe(
           'ShieldController:coverageResultReceived',
           (_coverageResult) => resolve(),
         );
       });
-      baseMessenger.publish(
+      rootMessenger.publish(
         'SignatureController:stateChange',
         {
           signatureRequests: {
@@ -324,15 +322,15 @@ describe('ShieldController', () => {
   });
 
   it('should check coverage for multiple signature request', async () => {
-    const { baseMessenger, backend } = setup();
+    const { rootMessenger, backend } = setup();
     const signatureRequest1 = generateMockSignatureRequest();
     const coverageResultReceived1 = new Promise<void>((resolve) => {
-      baseMessenger.subscribe(
+      rootMessenger.subscribe(
         'ShieldController:coverageResultReceived',
         (_coverageResult) => resolve(),
       );
     });
-    baseMessenger.publish(
+    rootMessenger.publish(
       'SignatureController:stateChange',
       {
         signatureRequests: {
@@ -348,12 +346,12 @@ describe('ShieldController', () => {
 
     const signatureRequest2 = generateMockSignatureRequest();
     const coverageResultReceived2 = new Promise<void>((resolve) => {
-      baseMessenger.subscribe(
+      rootMessenger.subscribe(
         'ShieldController:coverageResultReceived',
         (_coverageResult) => resolve(),
       );
     });
-    baseMessenger.publish(
+    rootMessenger.publish(
       'SignatureController:stateChange',
       {
         signatureRequests: {
@@ -384,7 +382,7 @@ describe('ShieldController', () => {
         updateSignatureRequest?: (signatureRequest: SignatureRequest) => void;
       },
     ) {
-      const { messenger, baseMessenger } = components;
+      const { messenger, rootMessenger } = components;
 
       // Create a promise that resolves when the state changes
       const stateUpdated = new Promise((resolve) =>
@@ -393,7 +391,7 @@ describe('ShieldController', () => {
 
       // Publish a signature request
       const signatureRequest = generateMockSignatureRequest();
-      baseMessenger.publish(
+      rootMessenger.publish(
         'SignatureController:stateChange',
         {
           signatureRequests: { [signatureRequest.id]: signatureRequest },
@@ -409,7 +407,7 @@ describe('ShieldController', () => {
       updatedSignatureRequest.status = SignatureRequestStatus.Signed;
       updatedSignatureRequest.rawSig = '0x00';
       options?.updateSignatureRequest?.(updatedSignatureRequest);
-      baseMessenger.publish(
+      rootMessenger.publish(
         'SignatureController:stateChange',
         {
           signatureRequests: { [signatureRequest.id]: updatedSignatureRequest },
@@ -478,7 +476,7 @@ describe('ShieldController', () => {
       components: ReturnType<typeof setup>,
       options?: { updateTransaction: (txMeta: TransactionMeta) => void },
     ) {
-      const { messenger, baseMessenger } = components;
+      const { messenger, rootMessenger } = components;
       // Create a promise that resolves when the state changes
       const stateUpdated = new Promise((resolve) =>
         messenger.subscribe('ShieldController:stateChange', resolve),
@@ -486,7 +484,7 @@ describe('ShieldController', () => {
 
       // Publish a transaction
       const txMeta = generateMockTxMeta();
-      baseMessenger.publish(
+      rootMessenger.publish(
         'TransactionController:stateChange',
         { transactions: [txMeta] } as TransactionControllerState,
         undefined as never,
@@ -500,7 +498,7 @@ describe('ShieldController', () => {
       updatedTxMeta.status = TransactionStatus.submitted;
       updatedTxMeta.hash = '0x00';
       options?.updateTransaction(updatedTxMeta);
-      baseMessenger.publish(
+      rootMessenger.publish(
         'TransactionController:stateChange',
         { transactions: [updatedTxMeta] } as TransactionControllerState,
         undefined as never,
@@ -559,7 +557,7 @@ describe('ShieldController', () => {
         deriveStateFromMetadata(
           controller.state,
           controller.metadata,
-          'anonymous',
+          'includeInDebugSnapshot',
         ),
       ).toMatchInlineSnapshot(`Object {}`);
     });