+++ packages/subscription-controller/src/SubscriptionController.test.ts
@@ -1,4 +1,11 @@
-import { deriveStateFromMetadata, Messenger } from '@metamask/base-controller';
+import { deriveStateFromMetadata } from '@metamask/base-controller';
+import {
+  Messenger,
+  MOCK_ANY_NAMESPACE,
+  type MessengerActions,
+  type MessengerEvents,
+  type MockAnyNamespace,
+} from '@metamask/messenger';
 import * as sinon from 'sinon';
 
 import {
@@ -9,7 +16,6 @@ import { SubscriptionServiceError } from './errors';
 import {
   getDefaultSubscriptionControllerState,
   SubscriptionController,
-  type AllowedActions,
   type AllowedEvents,
   type SubscriptionControllerMessenger,
   type SubscriptionControllerOptions,
@@ -25,6 +31,7 @@ import type {
   UpdatePaymentMethodOpts,
   Product,
   SubscriptionEligibility,
+  CachedLastSelectedPaymentMethods,
 } from './types';
 import {
   PAYMENT_TYPES,
@@ -34,6 +41,12 @@ import {
   SubscriptionUserEvent,
 } from './types';
 import { advanceTime } from '../../../tests/helpers';
+
+type AllActions = MessengerActions<SubscriptionControllerMessenger>;
+
+type AllEvents = MessengerEvents<SubscriptionControllerMessenger>;
+
+type RootMessenger = Messenger<MockAnyNamespace, AllActions, AllEvents>;
 
 // Mock data
 const MOCK_SUBSCRIPTION: Subscription = {
@@ -113,25 +126,30 @@ const MOCK_GET_SUBSCRIPTIONS_RESPONSE = {
 function createCustomSubscriptionMessenger(props?: {
   overrideEvents?: AllowedEvents['type'][];
 }) {
-  const baseMessenger = new Messenger<AllowedActions, AllowedEvents>();
+  const rootMessenger: RootMessenger = new Messenger({
+    namespace: MOCK_ANY_NAMESPACE,
+  });
 
-  const messenger = baseMessenger.getRestricted<
+  const messenger = new Messenger<
     typeof controllerName,
-    AllowedActions['type'],
-    AllowedEvents['type']
+    AllActions,
+    AllEvents,
+    RootMessenger
   >({
-    name: controllerName,
-    allowedActions: [
+    namespace: controllerName,
+    parent: rootMessenger,
+  });
+  rootMessenger.delegate({
+    messenger,
+    actions: [
       'AuthenticationController:getBearerToken',
       'AuthenticationController:performSignOut',
     ],
-    allowedEvents: props?.overrideEvents ?? [
-      'AuthenticationController:stateChange',
-    ],
+    events: props?.overrideEvents ?? ['AuthenticationController:stateChange'],
   });
 
   return {
-    baseMessenger,
+    rootMessenger,
     messenger,
   };
 }
@@ -140,25 +158,25 @@ function createCustomSubscriptionMessenger(props?: {
  * Jest Mock Utility to generate a mock Subscription Messenger
  *
  * @param overrideMessengers - override messengers if need to modify the underlying permissions
- * @param overrideMessengers.baseMessenger - base messenger to override
+ * @param overrideMessengers.rootMessenger - base messenger to override
  * @param overrideMessengers.messenger - messenger to override
  * @returns series of mocks to actions that can be called
  */
 function createMockSubscriptionMessenger(overrideMessengers?: {
-  baseMessenger: Messenger<AllowedActions, AllowedEvents>;
+  rootMessenger: RootMessenger;
   messenger: SubscriptionControllerMessenger;
 }) {
-  const { baseMessenger, messenger } =
+  const { rootMessenger, messenger } =
     overrideMessengers ?? createCustomSubscriptionMessenger();
 
   const mockPerformSignOut = jest.fn();
-  baseMessenger.registerActionHandler(
+  rootMessenger.registerActionHandler(
     'AuthenticationController:performSignOut',
     mockPerformSignOut,
   );
 
   return {
-    baseMessenger,
+    rootMessenger,
     messenger,
     mockPerformSignOut,
   };
@@ -216,7 +234,7 @@ type WithControllerCallback<ReturnValue> = (params: {
   controller: SubscriptionController;
   initialState: SubscriptionControllerState;
   messenger: SubscriptionControllerMessenger;
-  baseMessenger: Messenger<AllowedActions, AllowedEvents>;
+  rootMessenger: RootMessenger;
   mockService: ReturnType<typeof createMockSubscriptionService>['mockService'];
   mockPerformSignOut: jest.Mock;
 }) => Promise<ReturnValue> | ReturnValue;
@@ -237,7 +255,7 @@ async function withController<ReturnValue>(
   ...args: WithControllerArgs<ReturnValue>
 ) {
   const [{ ...rest }, fn] = args.length === 2 ? args : [{}, args[0]];
-  const { messenger, mockPerformSignOut, baseMessenger } =
+  const { messenger, mockPerformSignOut, rootMessenger } =
     createMockSubscriptionMessenger();
   const { mockService } = createMockSubscriptionService();
 
@@ -251,7 +269,7 @@ async function withController<ReturnValue>(
     controller,
     initialState: controller.state,
     messenger,
-    baseMessenger,
+    rootMessenger,
     mockService,
     mockPerformSignOut,
   });
@@ -1192,7 +1210,7 @@ describe('SubscriptionController', () => {
           deriveStateFromMetadata(
             controller.state,
             controller.metadata,
-            'anonymous',
+            'includeInDebugSnapshot',
           ),
         ).toMatchInlineSnapshot(`
           Object {
@@ -1401,6 +1419,74 @@ describe('SubscriptionController', () => {
             event: SubscriptionUserEvent.ShieldEntryModalViewed,
           }),
         ).rejects.toThrow(SubscriptionServiceError);
+      });
+    });
+  });
+
+  describe('cacheLastSelectedPaymentMethod', () => {
+    it('should cache last selected payment method successfully', async () => {
+      await withController(async ({ controller }) => {
+        controller.cacheLastSelectedPaymentMethod(PRODUCT_TYPES.SHIELD, {
+          type: PAYMENT_TYPES.byCard,
+          plan: RECURRING_INTERVALS.month,
+        });
+
+        expect(controller.state.lastSelectedPaymentMethod).toStrictEqual({
+          [PRODUCT_TYPES.SHIELD]: {
+            type: PAYMENT_TYPES.byCard,
+            plan: RECURRING_INTERVALS.month,
+          },
+        });
+      });
+    });
+
+    it('should update the last selected payment method for the same product', async () => {
+      await withController(
+        {
+          state: {
+            lastSelectedPaymentMethod: {
+              [PRODUCT_TYPES.SHIELD]: {
+                type: PAYMENT_TYPES.byCard,
+                plan: RECURRING_INTERVALS.month,
+              },
+            },
+          },
+        },
+        async ({ controller }) => {
+          expect(controller.state.lastSelectedPaymentMethod).toStrictEqual({
+            [PRODUCT_TYPES.SHIELD]: {
+              type: PAYMENT_TYPES.byCard,
+              plan: RECURRING_INTERVALS.month,
+            },
+          });
+
+          controller.cacheLastSelectedPaymentMethod(PRODUCT_TYPES.SHIELD, {
+            type: PAYMENT_TYPES.byCrypto,
+            paymentTokenAddress: '0x123',
+            plan: RECURRING_INTERVALS.month,
+          });
+
+          expect(controller.state.lastSelectedPaymentMethod).toStrictEqual({
+            [PRODUCT_TYPES.SHIELD]: {
+              type: PAYMENT_TYPES.byCrypto,
+              paymentTokenAddress: '0x123',
+              plan: RECURRING_INTERVALS.month,
+            },
+          });
+        },
+      );
+    });
+
+    it('should throw error when payment token address is not provided for crypto payment', async () => {
+      await withController(({ controller }) => {
+        expect(() =>
+          controller.cacheLastSelectedPaymentMethod(PRODUCT_TYPES.SHIELD, {
+            type: PAYMENT_TYPES.byCrypto,
+            plan: RECURRING_INTERVALS.month,
+          } as CachedLastSelectedPaymentMethods),
+        ).toThrow(
+          SubscriptionControllerErrorMessage.PaymentTokenAddressRequiredForCrypto,
+        );
       });
     });
   });